Description now pops up correctly.
Changed the layout of the search results page to make the popup
fit onto the page more organically. Removed much of the custom css
I had used and replaced it with Bootstrapiness.

Added a delay to the popover to give the description time to load
before displaying (having a description jump so drastically in length
was causing the popover arrow to appear in the wrong place on the
page).